Add pagination to grpc queries (#317)
* Start pagination for queries

* Complete pagination support

* Cleanups

* Use GRPC endpoint in CLI queries

* Decode page-key in cli

* Review feedback

+++ x/wasm/internal/keeper/test_fuzz.go
@@ -22,7 +22,7 @@ func FuzzAddrString(m *string, c fuzz.Continue) {
 }
 
 func FuzzAbsoluteTxPosition(m *types.AbsoluteTxPosition, c fuzz.Continue) {
-	m.BlockHeight = int64(c.RandUint64()) // can't be negative
+	m.BlockHeight = c.RandUint64()
 	m.TxIndex = c.RandUint64()
 }
